some fix in team detail
